Conflicts:
	Animal.java
	Organism.java

+++ Organism.java
@@ -10,7 +10,7 @@ public abstract class Organism {
         return location;
     }
 
-    public abstract int getCalories();
+    public abstract double getCalories();
     public abstract Image getImage();
     public abstract void step(Grid grid);
 }
